Cleaning up project structure
